requirements.txt

flask==3.0.0
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
pillow==10.1.0
qrcode==7.4.2
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2
matplotlib==3.8.2
werkzeug==3.0.1
